Vehicle usage.md
In ancient Greece, the hippeis were a class of aristocratic warriors who fought on horseback. This word is commonly found in Greek classical literature and is an important term for understanding the society and culture of the time. Coincidentally, the word "hippeis" is very similar to the word "hippie." For hippies, their vehicle, their mode of conduct, was love for their fellow human beings. Perhaps hippies belonged to the aristocratic class of the 20th century or perhaps not. I want to give a new approach to the concept of "aristocratic warrior" and have my horse, my vehicle, my mode of conduct, reflect through art the experiences of an aristocrat that was able to share love with their camarades traduce